backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.13.2
pydantic_core==2.46.2
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
